Implement remove downloaded JDK command (#3)
* Implement remove downloaded JDK command

This command has 2 modes, standard and force. Standard is the default,
and requires user-input to confirm deleting the selected directories is
okay. Force mode skips confirmation and can be done unattended.

Force mode can also be "faked" with this common bash-ism:
    yes | jpre remove <jdk_version>

If the JDK selected for removal is the JDK currently in use, then the
symlink will also be removed, to prevent broken symlinks.

* Rewrite some sections

Co-authored-by: Octavia Togami <octavia.togami@gmail.com>